requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.4.0
nltk==3.8.1
scikit-learn==1.5.0
sentence-transformers==3.0.1
python-dotenv==1.0.1
numpy==1.26.4
